Revert "円グラフの表示を実装"
