+++ .prettierignore
@@ -1,9 +1,3 @@
 # Ignore artifacts:
 build
 coverage
-
-# Ignore submodule
-data
-
-# Ignore shadcn/ui
-src/components/ui
