+++ components/ui/popover.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import * as React from "react"
 import * as PopoverPrimitive from "@radix-ui/react-popover"
