+++ components/ui/radio-group.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import * as React from "react"
 import * as RadioGroupPrimitive from "@radix-ui/react-radio-group"
 import { Circle } from "lucide-react"
