--- index.html
@@ -1,17 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" href="data:image/x-icon;," />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>福井県 観光オープンデータグラフ</title>
-    <meta
-      name="description"
-      content="Code4Fukuiらによって公開されているオープンデータから観光に関するデータの可視化を行います"
-    />
-  </head>
-  <body class="h-full w-full antialiased">
-    <div class="flex min-h-screen w-full flex-col items-center justify-center p-4" id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
