+++ interfaces/graph-series.interface.ts
@@ -1,18 +1,11 @@
-import { PartiallyRequired } from "@/lib/utils";
 import { OBJECT_CLASS, ObjectClass, ObjectClassAttribute } from "./aggregated-data.interface";
 import { Placement, PLACEMENTS } from "./placement.interface";
 
-// keyを変更すると後方互換性が失われるので注意
-export const GRAPH_TYPES = {
-  simple: "単純棒グラフ",
-  stack: "積み上げ棒グラフ",
-  ratio: "100%積み上げ棒グラフ",
-  pie: "円グラフ",
-} as const;
+export const GRAPH_TYPES = { simple: "単純", stack: "積み上げ", ratio: "100%積み上げ" } as const;
 export type GraphType = keyof typeof GRAPH_TYPES;
 
 export interface GraphSeries {
-  id?: string;
+  id: string;
   name?: string;
   show?: boolean;
   graphType: GraphType;
@@ -21,10 +14,6 @@ export interface GraphSeries {
   objectClass?: ObjectClass;
   exclude?: Record<string, string[]>;
 }
-
-export const isCartesian = (series: GraphSeries) =>
-  ["simpleBar", "stackBar", "ratioBar"].includes(series.graphType);
-export const isProportional = (series: GraphSeries) => ["simplePie"].includes(series.graphType);
 
 export const SERIES_PROPERTY_EFFECT_TO: Record<
   keyof GraphSeries,
@@ -48,14 +37,4 @@ export function defaultSeriesName(series: GraphSeries): string {
         ? "での"
         : ""
   }検出回数`;
-}
-
-export function isSeriesValid(
-  series: GraphSeries,
-): series is PartiallyRequired<GraphSeries, "placement" | "objectClass"> {
-  return (
-    series.placement !== undefined &&
-    series.objectClass !== undefined &&
-    (series.graphType === "simple" ? true : series.focusedAttribute !== undefined)
-  );
 }
