+++ interfaces/placement.interface.ts
