+++ package.json
@@ -1,20 +1,19 @@
 {
-  "name": "fukui-kanko-visualization-vite",
+  "name": "fukui-kanko-visualization",
+  "version": "0.1.0",
   "private": true,
-  "version": "0.0.0",
-  "type": "module",
   "scripts": {
     "submodule": "git submodule update --init --force --depth=1 --recursive && cp -R data/people-flow-data/full/* public",
-    "dev": "cp -R data/people-flow-data/full/* public && vite",
-    "build": "cp -R data/people-flow-data/full/* public && tsc -b && vite build",
-    "lint": "eslint .",
-    "preview": "vite preview",
+    "dev": "cp -R data/people-flow-data/full/* public && next dev",
+    "build": "cp -R data/people-flow-data/full/* public && next build",
+    "start": "pnpx serve@latest out",
+    "lint": "next lint",
     "prepare": "husky"
   },
   "lint-staged": {
-    "*.{ts,tsx,js,jsx}": [
+    "*.{ts,tsx}": [
       "prettier --write --ignore-unknown",
-      "eslint --cache --fix"
+      "next lint --fix"
     ],
     "**/*": "prettier --write --ignore-unknown"
   },
@@ -22,45 +21,42 @@
     "@primer/octicons-react": "^19.14.0",
     "@radix-ui/react-checkbox": "^1.1.3",
     "@radix-ui/react-dialog": "^1.1.5",
+    "@radix-ui/react-hover-card": "^1.1.4",
     "@radix-ui/react-label": "^2.1.1",
-    "@radix-ui/react-popover": "^1.1.5",
+    "@radix-ui/react-popover": "^1.1.4",
     "@radix-ui/react-radio-group": "^1.2.2",
-    "@radix-ui/react-select": "^2.1.5",
+    "@radix-ui/react-select": "^2.1.4",
     "@radix-ui/react-slot": "^1.1.1",
+    "@radix-ui/react-tooltip": "^1.1.6",
+    "@types/papaparse": "^5.3.15",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
-    "lucide-react": "^0.474.0",
-    "papaparse": "^5.5.2",
-    "react": "^18.3.1",
+    "lucide-react": "^0.471.1",
+    "next": "^15.1.4",
+    "papaparse": "^5.5.1",
+    "react": "19.0.0-rc-02c0e824-20241028",
+    "react-apexcharts": "^1.7.0",
     "react-day-picker": "8.10.1",
-    "react-dom": "^18.3.1",
-    "recharts": "^2.15.1",
-    "tailwind-merge": "^3.0.1",
+    "react-dom": "19.0.0-rc-02c0e824-20241028",
+    "recharts": "^2.15.0",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@eslint/js": "^9.17.0",
-    "@trivago/prettier-plugin-sort-imports": "^5.2.2",
-    "@types/node": "^22.12.0",
-    "@types/papaparse": "^5.3.15",
+    "@types/node": "^20.17.13",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
-    "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.17.0",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.16",
-    "globals": "^15.14.0",
+    "eslint": "^8.57.1",
+    "eslint-config-next": "15.0.2",
     "husky": "^9.1.7",
-    "lint-staged": "^15.4.3",
+    "lint-staged": "^15.3.0",
     "postcss": "^8.5.1",
-    "prettier": "^3.4.2",
-    "prettier-plugin-tailwindcss": "^0.6.11",
+    "prettier": "3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.9",
     "tailwindcss": "^3.4.17",
-    "typescript": "~5.6.2",
-    "typescript-eslint": "^8.18.2",
-    "vite": "^6.0.5"
+    "typescript": "^5.7.3"
   },
   "packageManager": "pnpm@9.10.0+sha512.73a29afa36a0d092ece5271de5177ecbf8318d454ecd701343131b8ebc0c1a91c487da46ab77c8e596d6acf1461e3594ced4becedf8921b074fbd8653ed7051c",
   "pnpm": {
