+++ tailwind.config.ts
@@ -1,0 +1,100 @@
+import type { Config } from "tailwindcss";
+import { PluginCreator } from "tailwindcss/types/config";
+import tailwindAnimate from "tailwindcss-animate";
+
+const config: Config = {
+  darkMode: ["class"],
+  content: [
+    "./pages/**/*.{js,ts,jsx,tsx,mdx}",
+    "./components/**/*.{js,ts,jsx,tsx,mdx}",
+    "./app/**/*.{js,ts,jsx,tsx,mdx}",
+  ],
+  theme: {
+    extend: {
+      colors: {
+        primary: {
+          DEFAULT: "hsl(var(--primary))",
+          foreground: "hsl(var(--primary-foreground))",
+        },
+        background: "hsl(var(--background))",
+        foreground: "hsl(var(--foreground))",
+        card: {
+          DEFAULT: "hsl(var(--card))",
+          foreground: "hsl(var(--card-foreground))",
+        },
+        popover: {
+          DEFAULT: "hsl(var(--popover))",
+          foreground: "hsl(var(--popover-foreground))",
+        },
+        secondary: {
+          DEFAULT: "hsl(var(--secondary))",
+          foreground: "hsl(var(--secondary-foreground))",
+        },
+        muted: {
+          DEFAULT: "hsl(var(--muted))",
+          foreground: "hsl(var(--muted-foreground))",
+        },
+        accent: {
+          DEFAULT: "hsl(var(--accent))",
+          foreground: "hsl(var(--accent-foreground))",
+        },
+        destructive: {
+          DEFAULT: "hsl(var(--destructive))",
+          foreground: "hsl(var(--destructive-foreground))",
+        },
+        border: "hsl(var(--border))",
+        input: "hsl(var(--input))",
+        ring: "hsl(var(--ring))",
+        star: "hsl(var(--star))",
+        chart: {
+          "1": "hsl(var(--chart-1))",
+          "2": "hsl(var(--chart-2))",
+          "3": "hsl(var(--chart-3))",
+          "4": "hsl(var(--chart-4))",
+          "5": "hsl(var(--chart-5))",
+        },
+      },
+      gridTemplateColumns: {
+        edgetip: "64px 1fr 64px",
+      },
+      borderRadius: {
+        lg: "var(--radius)",
+        md: "calc(var(--radius) - 2px)",
+        sm: "calc(var(--radius) - 4px)",
+      },
+    },
+  },
+  plugins: [
+    ((api) => {
+      api.addUtilities({
+        ".tooltip": {
+          position: "relative",
+          display: "inline-block",
+          cursor: "pointer",
+          "&::before": {
+            content: "attr(data-tooltip)",
+            position: "absolute",
+            bottom: "100%",
+            left: "50%",
+            transform: "translateX(-50%)",
+            "background-color": "#333",
+            color: "white",
+            padding: "5px",
+            "border-radius": "5px",
+            "font-size": "12px",
+            "white-space": "nowrap",
+            opacity: "0",
+            visibility: "hidden",
+            transition: "opacity 0.3s",
+          },
+          "&:hover::before": {
+            opacity: "1",
+            visibility: "visible",
+          },
+        },
+      });
+    }) as PluginCreator,
+    tailwindAnimate,
+  ],
+};
+export default config;
