--- vite.config.ts
@@ -1,17 +1,0 @@
-import react from "@vitejs/plugin-react";
-import path from "path";
-import { defineConfig } from "vite";
-
-// https://vite.dev/config/
-export default defineConfig({
-  base: process.env.PAGES ? "fukui-kanko-people-flow-visualization" : "./",
-  plugins: [react()],
-  resolve: {
-    alias: {
-      "@": path.resolve(__dirname, "./src"),
-    },
-  },
-  server: {
-    port: 3000,
-  },
-});
